created coords and basic test file for coords
